Refine current map invalidation to invalidate only if the new URL has a mapname parameter and it's current mapname differs
NOTE: This is mainly for task pane content that POSTs into itself and in such cases the mapname is lost in the URL. This will also cause said content to not invalidate if the active map changes. To work around this, there will need to be client-side checks that the current map name is still the same.

Fixes #186