Add Bing Maps support.
NOTE: If loading from an Application Definition, you will need to put your Bing Maps API key in the top-level <BingMapKey> extension property of the Application Definition document. This viewer will look for this key and require it in order to add Bing Maps layers.

Maestro will be updated to be able to edit this (see: https://github.com/jumpinjackie/mapguide-maestro/issues/38)

Fixes #189