Move flyout open/closed state out of component state that is local to flyout/component menu items and have this state passed down from the redux state tree (where this state was already been stored, but not passed down to components that needed to know).
Basically, this change allows for flyout menus to behave more intuitively. Clicking an item inside, flows the flyout closed state to the redux store (which it already did), which then flows back down to the affected flyout menu items (which it didn't do until now). So not only is the flyout menu "visually" closed, it is also closed state-wise.

Fixes #290