#318: First part of hopefully allowing for custom reducers to be added to a custom build of this viewer.
src/reducers/index no longer returns a combined reducer, instead it returns the argument that would be passed to combineReducers().

The one responsible for calling combineReducers is the ApplicationViewModel, which has overridable hooks to inject extra initial state and custom reducers.

Thus the means of adding custom reducers is to sub-class ApplicationViewModel and provide the custom reducers and initial state there. Then in the custom build's entry point your ApplicationViewModel's subclass is exported in place of (or in addition to), the ApplicationViewModel.

At least that's the theory. Need to make a npm module with this change to test it out.