Port across our own copy of ol.source.ImageMapGuide
It turns out we are indeed able to create our own custom OL sources by using ol.inherits().

The only catch is that because we are working in a TypeScript environment, we have to port it across in its raw prototype-based style instead of the more naturally intuitive TS way of defining classes and inheritance as I cannot verify whether implementing this MapGuide source as a regular TS class can be used with the ol.inherits() mechanism of inheritance.

Fixes #379