Separate out css to an external stylesheet via mini-css-extract-plugin
This was absolute hell due to different loaders interpreting paths and URLs of certain static assets differently.

After painstaking investigation, the workaround that was most acceptable was to move the image sprite out to a separate directory and add special webpack loader rules just for handling this specific asset.

With this change the JS bundle is now pure JS, giving us a better view of actual bundle size.

And because CSS is now externalized, users can directly modify styles to their liking.

Fixes #645