Allow custom icons to be injected into a layer or group entry of a Legend component.
A viewer may be mounted with a custom legend.provideExtraLayerIconsHtml or legend. provideExtraGroupIconsHtml option which specifies a function to add extra HTML elements for a given layer or group.

The intention behind these custom icons is to serve as links for layer or group-specific functionality (eg. A download data link for a layer)

Fixes #1441